services/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sima-api"
version = "0.1.0"
description = "SIMA API - Backend for web frontend"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
keywords = ["sima", "api", "backend", "websocket"]
authors = [
    { name = "SIMA Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.8.0",
    "sima-core",
    "sima-storage",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
sima-api = "sima_api.main:run"

[tool.hatch.build.targets.wheel]
packages = ["sima_api"]
